requirements.txt
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
wincertstore==0.2